+++ src/mock/filter.js
@@ -1,4 +1,4 @@
-import {isTaskRepeating, getDaysToDeadline} from "../utils.js";
+import {isTaskRepeating, getDaysToDeadline} from "../utils/task.js";
 
 const taskToFilterMap = {
   all: (tasks) => tasks.filter((task) => !task.isArchive).length,
