+++ src/utils/common.js
@@ -1,0 +1,6 @@
+export const getRandomInteger = (a, b) => {
+  const lower = Math.ceil(Math.min(a, b));
+  const upper = Math.floor(Math.max(a, b));
+
+  return Math.round(lower + Math.random() * (upper - lower));
+};
