+++ src/view/task.js
@@ -1,5 +1,5 @@
 import AbstractView from "./abstract.js";
-import {getDaysToDeadline, isTaskRepeating, humanizeTaskDueDate} from "../utils.js";
+import {getDaysToDeadline, isTaskRepeating, humanizeTaskDueDate} from "../utils/task.js";
 
 const createTaskTemplate = (task) => {
   const {color, description, dueDate, repeating, isArchive, isFavorite} = task;
